move categoryFilter to components folder
